Remove rapid-fire F12
When pressing NAV+# on boot,
I can never be sure whether the NAV layer is active
(whether the NAV modifier has been pressed before or after
the keyboard is turned on), so it's not that useful in the end.

In addition to that, using F12 in applications -
to toggle the developer tools in Firefox
and to toggle a panel in MuseScore - becomes cumbersome,
as the opened panels flicker due to F12 getting pressed multiple times.

Also, there is a reliable way to open the boot menu on ThinkPads -
disable quick boot in UEFI setup.
Then, as soon as the POST screen appears, keypresses from the keyboard
are recognized immediately, without any delays.